Move DeviceSchedule to seperate model file
